+++ src/Gate.java
@@ -12,7 +12,7 @@ public class Gate extends BasicOperations{
     @Override
     public void draw(Graphics graphics) {
         graphics.setColor(new Color(80, 50, 25));
-        graphics.fillRect(dataset.x, dataset.y, dataset.sizex, dataset.sizey);
+        graphics.fillRect((int)dataset.x, (int)dataset.y, (int)dataset.sizex, (int)dataset.sizey);
     }
 
     @Override
