+++ src/Obstacle.java
@@ -12,7 +12,7 @@ public class Obstacle extends BasicOperations{
     @Override
     public void draw(Graphics graphics) {
         graphics.setColor(Color.darkGray);
-        graphics.fillRect(dataset.x, dataset.y, dataset.sizex, dataset.sizey);
+        graphics.fillRect((int)dataset.x, (int)dataset.y, (int)dataset.sizex, (int)dataset.sizey);
     }
 
     @Override
